split implementation into several files for readability
